Server/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
anyio==3.6.2
argilla==1.7.0
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
bleach==6.0.0
boto3==1.26.142
botocore==1.29.142
bs4==0.0.1
casbin==1.19.0
casbin-sqlalchemy-adapter==0.5.2
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
click-log==0.4.0
colorama==0.4.6
commonmark==0.9.1
cryptography==40.0.1
dataclasses-json==0.5.7
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.1.0
docutils==0.20.1
dotty-dict==1.3.1
ecdsa==0.18.0
et-xmlfile==1.1.0
fastapi==0.95.1
fastapi-pagination==0.12.3
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
gotrue==1.0.1
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
importlib-metadata==6.6.0
invoke==1.7.3
jaraco.classes==3.2.3
jmespath==1.0.1
joblib==1.2.0
jose==1.0.0
keyring==23.13.1
langchain==0.0.212
langchainplus-sdk==0.0.17
lxml==4.9.2
Mako==1.2.4
Markdown==3.4.3
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
monotonic==1.6
more-itertools==9.1.0
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
mysql-connector-python==8.0.33
ndg-httpsclient==0.5.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.23.5
oauthlib==3.2.2
olefile==0.46
openai==0.27.7
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
packaging==23.1
pandas==1.5.3
passlib==1.7.4
pdf2image==1.16.3
pdfminer.six==20221105
Pillow==9.5.0
pkginfo==1.9.6
postgrest==0.10.6
protobuf==3.20.3
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.7
Pygments==2.15.1
PyJWT==2.6.0
PyMySQL==1.0.3
pyOpenSSL==23.1.1
pypandoc==1.11
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-gitlab==3.14.0
python-jose==3.3.0
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
python-semantic-release==7.33.2
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
readme-renderer==37.3
realtime==1.0.0
regex==2023.5.5
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==13.0.1
rsa==4.9
s3transfer==0.6.1
semver==2.13.0
simpleeval==0.9.13
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
social-auth-core==4.4.2
soupsieve==2.4.1
SQLAlchemy==2.0.12
starlette==0.26.1
storage3==0.5.2
StrEnum==0.4.10
supabase==1.0.3
supafunc==0.2.2
tabulate==0.9.0
tenacity==8.2.2
tiktoken==0.4.0
tomlkit==0.11.8
tqdm==4.65.0
twine==3.8.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.5.0
tzdata==2023.3
unstructured==0.6.10
urllib3==1.26.15
uvicorn==0.15.0
uvloop==0.17.0
webencodings==0.5.1
websockets==10.4
wrapt==1.14.1
xlrd==2.0.1
XlsxWriter==3.1.2
yarl==1.9.2
zipp==3.15.0
